requirements.txt
openai
google-genai
faiss-cpu
sentence-transformers
numpy
python-dotenv
pypdf
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg[binary]
alembic
tenacity
axios
